requirements.txt
shapely
requests
dateutils
google-genai
dotenv
